Move contact congress interactions to lib
