+++ app/models/congress_message_campaign.rb
@@ -1,4 +1,3 @@
-require "rest_client"
 class CongressMessageCampaign < ActiveRecord::Base
   belongs_to :topic_category
   has_one :action_page
@@ -32,22 +31,21 @@ class CongressMessageCampaign < ActiveRecord::Base
   end
 
   def date_fills(start_date = nil, end_date = nil)
-    begin
-      r = RestClient.get(date_fills_url(start_date, end_date))
-      JSON.parse r
-    rescue RestClient::ExceptionWithResponse => e
-      Rails.logger.error e
-      return {}
-    end
-  end
-
-  def date_fills_url(start_date = nil, end_date = nil, bioguide_id = nil)
     params = {
       date_start: start_date,
       date_end: end_date,
       campaign_tag: campaign_tag
     }.compact
-    CongressMessageCampaign.url("/successful-fills-by-date/", params, bioguide_id)
+    CongressForms.post("/successful-fills-by-date/", params)
+  end
+
+  def date_fills_url(start_date = nil, end_date = nil)
+    params = {
+      date_start: start_date,
+      date_end: end_date,
+      campaign_tag: campaign_tag
+    }.compact
+    CongressForms.url("/successful-fills-by-date/", params)
   end
 
   private
@@ -67,13 +65,5 @@ class CongressMessageCampaign < ActiveRecord::Base
     self.alt_text_extra_fields_explain = nil if alt_text_extra_fields_explain.blank?
     self.alt_text_look_up_helper = nil if alt_text_look_up_helper.blank?
     self.alt_text_customize_message_helper = nil if alt_text_customize_message_helper.blank?
-  end
-
-  def self.url(path = "/", params = {}, bioguide_id = nil)
-    url = Rails.application.config.congress_forms_url + path
-    url += bioguide_id unless bioguide_id.nil?
-    url += "?" + {
-      debug_key: Rails.application.secrets.congress_forms_debug_key,
-    }.merge(params).to_query
   end
 end
